requirements.txt

#Core libs
django    				#core django libs
django-localflavor
dj_static


#REST libraries
djangorestframework 		  #REST framework
markdown
django-filter
django-oauth2-provider        #OAuth2 provider
requests					  #HTTP client

#DB
mysql-python
